move ViewPrivate into separate file
Conflicts:
	packages/perspective-viewer/src/js/view.js
	packages/perspective/src/config/common.config.js
