+++ packages/perspective-viewer/README.md
@@ -2,44 +2,29 @@
 
 ### Table of Contents
 
--   [registerPlugin][1]
--   [View][2]
-    -   [sort][3]
-    -   [columns][4]
-    -   [computed-columns][5]
-    -   [aggregates][6]
-    -   [filters][7]
+-   [View][1]
+    -   [sort][2]
+    -   [columns][3]
+    -   [computed-columns][4]
+    -   [aggregates][5]
+    -   [filters][6]
+    -   [view][7]
     -   [view][8]
-    -   [view][9]
-    -   [column-pivots][10]
-    -   [row-pivots][11]
-    -   [message][12]
-    -   [worker][13]
-    -   [load][14]
-    -   [update][15]
-    -   [notifyResize][16]
-    -   [copy][17]
-    -   [delete][18]
-    -   [save][19]
-    -   [restore][20]
-    -   [reset][21]
-    -   [handleClipboardCopy][22]
+    -   [column-pivots][9]
+    -   [row-pivots][10]
+    -   [message][11]
+    -   [worker][12]
+    -   [load][13]
+    -   [update][14]
+    -   [notifyResize][15]
+    -   [copy][16]
+    -   [delete][17]
+    -   [save][18]
+    -   [restore][19]
+    -   [reset][20]
+    -   [handleClipboardCopy][21]
+-   [View#perspective-view-update][22]
 -   [View#perspective-config-update][23]
--   [View#perspective-view-update][24]
-
-## registerPlugin
-
-Register a plugin with the <perspective-viewer> component.
-
-**Parameters**
-
--   `name` **[string][25]** The logical unique name of the plugin.  This will be
-    used to set the component's `view` attribute.
--   `plugin` **[object][26]** An object with this plugin's prototype.  Valid keys are:
-        name : The display name for this plugin.
-        create (required) : The creation function - may return a `Promise`.
-        delete : The deletion function.
-        mode : The selection mode - may be "toggle" or "select".
 
 ## View
 
@@ -73,7 +58,7 @@ The set of visible columns.
 
 **Parameters**
 
--   `columns` **[array][27]** An array of strings, the names of visible columns.
+-   `columns` **[array][24]** An array of strings, the names of visible columns.
 
 **Examples**
 
@@ -96,7 +81,7 @@ The set of visible columns.
 
 **Parameters**
 
--   `computed-columns` **[array][27]** An array of computed column objects
+-   `computed-columns` **[array][24]** An array of computed column objects
 
 **Examples**
 
@@ -119,11 +104,11 @@ The set of column aggregate configurations.
 
 **Parameters**
 
--   `aggregates` **[object][26]** A dictionary whose keys are column names, and
+-   `aggregates` **[object][25]** A dictionary whose keys are column names, and
     values are valid aggregations.  The `aggergates` attribute works as an
     override;  in lieu of a key for a column supplied by the developers, a
     default will be selected and reflected to the attribute based on the
-    column's type.  See [perspective/src/js/defaults.js][28]
+    column's type.  See [perspective/src/js/defaults.js][26]
 
 **Examples**
 
@@ -192,7 +177,7 @@ render it hidden until the message is removed.
 
 **Parameters**
 
--   `msg` **[string][25]** The message. This can be HTML - it is not sanitized.
+-   `msg` **[string][27]** The message. This can be HTML - it is not sanitized.
 
 **Examples**
 
@@ -254,7 +239,7 @@ const tbl = perspective.table("x,y\n1,a\n2,b");
 my_viewer.load(tbl);
 ```
 
-Returns **[Promise][29]&lt;void>** A promise which resolves once the data is
+Returns **[Promise][28]&lt;void>** A promise which resolves once the data is
 loaded and a `perspective.view` has been created.
 
 ### update
@@ -296,14 +281,14 @@ Deletes this element's data and clears it's internal state (but not its
 user state).  This (or the underlying `perspective.table`'s equivalent
 method) must be called in order for its memory to be reclaimed.
 
-Returns **[Promise][29]&lt;[boolean][30]>** Whether or not this call resulted in the
+Returns **[Promise][28]&lt;[boolean][29]>** Whether or not this call resulted in the
 underlying `perspective.table` actually being deleted.
 
 ### save
 
 Serialize this element's attribute/interaction state.
 
-Returns **[object][26]** a serialized element.
+Returns **[object][25]** a serialized element.
 
 ### restore
 
@@ -312,9 +297,9 @@ Restore this element to a state as generated by a reciprocal call to
 
 **Parameters**
 
--   `x` **[object][26]** returned by `save`.
+-   `x` **[object][25]** returned by `save`.
 
-Returns **[Promise][29]&lt;void>** A promise which resolves when the changes have
+Returns **[Promise][28]&lt;void>** A promise which resolves when the changes have
 been applied.
 
 ### reset
@@ -328,80 +313,78 @@ state.
 Copies this element's view data (as a CSV) to the clipboard.  This method
 must be called from an event handler, subject to the browser's
 restrictions on clipboard access.  See
-[https://www.w3.org/TR/clipboard-apis/#allow-read-clipboard][31].
+[https://www.w3.org/TR/clipboard-apis/#allow-read-clipboard][30].
 
 **Parameters**
 
 -   `options`  
-
-## View#perspective-config-update
-
-`perspective-config-update` is fired whenever an configuration attribute has
-been modified, by the user or otherwise.
 
 ## View#perspective-view-update
 
 `perspective-view-update` is fired whenever underlying `view`'s data has
 updated, including every invocation of `load` and `update`.
 
-[1]: #registerplugin
+## View#perspective-config-update
 
-[2]: #view
+`perspective-config-update` is fired whenever an configuration attribute has
+been modified, by the user or otherwise.
 
-[3]: #sort
+[1]: #view
 
-[4]: #columns
+[2]: #sort
 
-[5]: #computed-columns
+[3]: #columns
 
-[6]: #aggregates
+[4]: #computed-columns
 
-[7]: #filters
+[5]: #aggregates
 
-[8]: #view-1
+[6]: #filters
 
-[9]: #view-2
+[7]: #view-1
 
-[10]: #column-pivots
+[8]: #view-2
 
-[11]: #row-pivots
+[9]: #column-pivots
 
-[12]: #message
+[10]: #row-pivots
 
-[13]: #worker
+[11]: #message
 
-[14]: #load
+[12]: #worker
 
-[15]: #update
+[13]: #load
 
-[16]: #notifyresize
+[14]: #update
 
-[17]: #copy
+[15]: #notifyresize
 
-[18]: #delete
+[16]: #copy
 
-[19]: #save
+[17]: #delete
 
-[20]: #restore
+[18]: #save
 
-[21]: #reset
+[19]: #restore
 
-[22]: #handleclipboardcopy
+[20]: #reset
+
+[21]: #handleclipboardcopy
+
+[22]: #viewperspective-view-update
 
 [23]: #viewperspective-config-update
 
-[24]: #viewperspective-view-update
+[24]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array
 
-[25]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String
+[25]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Object
 
-[26]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Object
+[26]: perspective/src/js/defaults.js
 
-[27]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array
+[27]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String
 
-[28]: perspective/src/js/defaults.js
+[28]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Promise
 
-[29]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Promise
+[29]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Boolean
 
-[30]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Boolean
-
-[31]: https://www.w3.org/TR/clipboard-apis/#allow-read-clipboard
+[30]: https://www.w3.org/TR/clipboard-apis/#allow-read-clipboard
