+++ packages/perspective-viewer/src/js/view/actions.js
@@ -1,6 +1,6 @@
 /******************************************************************************
  *
- * Copyright (c) 2017, the Perspective Authors.
+ * Copyright (c) 2018, the Perspective Authors.
  *
  * This file is part of the Perspective library, distributed under the terms of
  * the Apache License 2.0.  The full license can be found in the LICENSE file.
